--- a/src/Core/Framework/DataAbstractionLayer/Field/TranslatedField.php
+++ b/src/Core/Framework/DataAbstractionLayer/Field/TranslatedField.php
@@ -16,7 +16,10 @@ class TranslatedField extends Field
 
     private readonly string $foreignFieldName;
 
-    public function __construct(string $propertyName)
+    /**
+     * @param bool $prefilledFallback - only relevant in OpenSearch context, if true, the translated field is filled with fallback value if no translation is available.
+     */
+    public function __construct(string $propertyName, private readonly bool $prefilledFallback = false)
     {
         $this->foreignClassName = LanguageDefinition::class;
         $this->foreignFieldName = 'id';
@@ -37,6 +40,11 @@ class TranslatedField extends Field
     public function getForeignFieldName(): string
     {
         return $this->foreignFieldName;
+    }
+
+    public function isPrefilledFallback(): bool
+    {
+        return $this->prefilledFallback;
     }
 
     protected function getSerializerClass(): string