--- a/src/Elasticsearch/Framework/Command/ElasticsearchIndexingCommand.php
+++ b/src/Elasticsearch/Framework/Command/ElasticsearchIndexingCommand.php
@@ -80,6 +80,10 @@ class ElasticsearchIndexingCommand extends Command
 
             $this->messageBus->dispatch($message);
 
+            if ($message->isLastMessage()) {
+                break;
+            }
+
             $progressBar->advance($step);
         }