--- a/src/Elasticsearch/Framework/Command/ElasticsearchIndexingCommand.php
+++ b/src/Elasticsearch/Framework/Command/ElasticsearchIndexingCommand.php
@@ -7,6 +7,7 @@ use Shopware\Core\Framework\DataAbstractionLayer\Command\ConsoleProgressTrait;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Elasticsearch\Framework\Indexing\CreateAliasTaskHandler;
 use Shopware\Elasticsearch\Framework\Indexing\ElasticsearchIndexer;
+use Shopware\Elasticsearch\Framework\Indexing\ElasticsearchIndexingMessage;
 use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Helper\ProgressBar;
@@ -65,9 +66,25 @@ class ElasticsearchIndexingCommand extends Command
 
         $entities = $input->getOption('only') ? explode(',', $input->getOption('only')) : [];
         $offset = null;
+        $messagesToDispatch = [];
+
         while ($message = $this->indexer->iterate($offset, $entities)) {
             $offset = $message->getOffset();
 
+            $messagesToDispatch[] = $message;
+        }
+
+        $lastMessage = end($messagesToDispatch);
+
+        if (!$lastMessage instanceof ElasticsearchIndexingMessage) {
+            $this->io->error('No messages found for indexing.');
+
+            return self::SUCCESS;
+        }
+
+        $lastMessage->markAsLastMessage();
+
+        foreach ($messagesToDispatch as $message) {
             $step = \count($message->getData()->getIds());
 
             if ($input->getOption('no-queue')) {
@@ -79,10 +96,6 @@ class ElasticsearchIndexingCommand extends Command
             }
 
             $this->messageBus->dispatch($message);
-
-            if ($message->isLastMessage()) {
-                break;
-            }
 
             $progressBar->advance($step);
         }