--- a/src/Elasticsearch/Framework/Indexing/ElasticsearchIndexingMessage.php
+++ b/src/Elasticsearch/Framework/Indexing/ElasticsearchIndexingMessage.php
@@ -17,7 +17,8 @@ class ElasticsearchIndexingMessage implements AsyncMessageInterface, Deduplicata
     public function __construct(
         private readonly IndexingDto $data,
         private readonly ?IndexerOffset $offset,
-        private readonly Context $context
+        private readonly Context $context,
+        private readonly bool $lastMessage = false
     ) {
     }
 
@@ -53,5 +54,10 @@ class ElasticsearchIndexingMessage implements AsyncMessageInterface, Deduplicata
         ]);
 
         return Hasher::hash($data);
+    }
+
+    public function isLastMessage(): bool
+    {
+        return $this->lastMessage;
     }
 }