--- a/src/Elasticsearch/Framework/Indexing/Event/ElasticsearchIndexingFinished.php
+++ b/src/Elasticsearch/Framework/Indexing/Event/ElasticsearchIndexingFinished.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Elasticsearch\Framework\Indexing\Event;
+
+use Shopware\Core\Framework\Log\Package;
+
+/**
+ * @codeCoverageIgnore
+ */
+#[Package('framework')]
+class ElasticsearchIndexingFinished
+{
+}