… changelog/_unreleased/2025-08-12-optimize-storefront-elasticsearch.md
… src/Elasticsearch/Framework/Indexing/ElasticsearchIndexer.php
… src/Elasticsearch/Framework/Indexing/Event/ElasticsearchIndexingFinishedEvent.php
… src/Elasticsearch/Framework/SystemUpdateListener.php
… src/Elasticsearch/Product/ElasticsearchOptimizeSwitch.php
… src/Elasticsearch/Product/ElasticsearchProductDefinition.php
… src/Elasticsearch/Product/ProductCriteriaParser.php
… src/Elasticsearch/Resources/config/packages/elasticsearch.yaml
… src/Elasticsearch/Resources/config/services.xml
… src/Elasticsearch/TokenQueryBuilder.php
… tests/integration/Elasticsearch/Product/ElasticsearchProductTest.php
… tests/integration/Elasticsearch/Product/Fixture/ProductsFixture.php
… tests/integration/Elasticsearch/Product/ProductSearchQueryBuilderTest.php
… tests/migration/Core/V6_8/Migration1743256470RemoveElasticsearchAppConfigFlagTest.php
… tests/unit/Core/Content/Product/SalesChannel/ProductAvailableFilterTest.php
… tests/unit/Core/Framework/DataAbstractionLayer/Dbal/EntityDefinitionQueryHelperTest.php
… tests/unit/Core/Framework/DataAbstractionLayer/Field/TranslatedFieldTest.php
… tests/unit/Core/Framework/DataAbstractionLayer/Search/Filter/NotEqualsFilterTest.php
… tests/unit/Elasticsearch/Framework/Command/ElasticsearchIndexingCommandTest.php
… tests/unit/Elasticsearch/Framework/DataAbstractionLayer/CriteriaParserTest.php
… tests/unit/Elasticsearch/Framework/DataAbstractionLayer/ElasticsearchEntitySearcherTest.php
… tests/unit/Elasticsearch/Framework/Indexing/ElasticsearchIndexerTest.php
… tests/unit/Elasticsearch/Product/ElasticsearchOptimizeSwitchTest.php
… tests/unit/Elasticsearch/Product/ElasticsearchProductDefinitionTest.php
… tests/unit/Elasticsearch/Product/ProductCriteriaParserTest.php
… tests/unit/Elasticsearch/Product/ProductSearchQueryBuilderTest.php
… tests/unit/Elasticsearch/TokenQueryBuilderTest.php