### Prerequisites

Number of documents: 2580421
Number of sales channel languages: 5
Endpoint: /sw_product/_search?request_cache=false

Queries using for testing

1. With a sort script

```
{
  "profile": true,
  "_source": false,
  "sort": [
    {
      "_script": {
        "type": "string",
        "script": {
          "source": "def languages = params['languages'];\ndef suffix = params.containsKey('suffix') ? '.' + params['suffix'] : '';\n\nfor (int i = 0; i < languages.length; i++) {\n    def field_name = params['field'] + '.' + languages[i] + suffix;\n\n    if (doc[field_name].size() > 0 && doc[field_name].value != null && doc[field_name].value.toString().length() > 0) {\n        def fieldValue = doc[field_name].value;\n\n        return fieldValue.toString();\n    }\n}\n\nreturn '';\n",
          "lang": "painless",
          "params": {
            "field": "name",
            "languages": [
              "0198187d773b7833806c564766c15e4c",
              "2fbb5fe2e29a4d70aa5854ce7ce3e20b"
            ]
          }
        },
        "order": "DESC"
      }
    },
    {
      "id": {
        "order": "ASC"
      }
    }
  ],
  "aggregations": {
    "total-count": {
      "cardinality": {
        "field": "displayGroup"
      }
    }
  },
  "from": 0,
  "size": 24,
  "collapse": {
    "field": "displayGroup"
  },
  "timeout": "5s"
}
```

2. Without sort script

```
{
  "profile": true,
  "_source": false,
  "sort": [
    {
      "name.0198187d773b7833806c564766c15e4c": {
        "order": "DESC"
      }
    },
    {
      "id": {
        "order": "ASC"
      }
    }
  ],
  "aggregations": {
    "total-count": {
      "cardinality": {
        "field": "displayGroup"
      }
    }
  },
  "from": 0,
  "size": 24,
  "collapse": {
    "field": "displayGroup"
  },
  "timeout": "5s"
}
```

### Benchmark result:

|Criteria|Before (with script sort)|After (without script sort)|
| -------- | ------- |------- |
|Index time (no-queue) | 2 mins, 56 secs | 2 mins, 57 secs |
|Disk space | 366 MB | 433 MB |
|Search time | ~150-300ms | ~20-30ms |
|Cost for sorting(from the profile result)|~130ms (85%)|~6.5ms (22%)|

### Conclusion

- Search time dropped from ~150–300ms to ~20–30ms, a 5–10x improvement, despite identical data size, indexing time and index structure.

- The collector time related to sorting, as reported in the query profile, decreased from ~130ms (85% of total query time) to just ~6.5ms (22%). This confirms that the script-based sort was the dominant bottleneck.

- An increase in disk usage (366MB → 433MB) is the result of storing sortable content in 5 different languages

So, for large-scale catalogs (2M+ products), avoid script sort and use pre-indexed sortable fields wherever possible. This is a highly recommended best practice. The trade-off in storage is minimal and well worth the massive gains in query responsiveness.